refactor load usfm
